requirements.txt
aiogram>=3.17.0
aiohttp>=3.11.0
requests>=2.32.0
python-dotenv>=1.0.0
uvloop>=0.21.0
psycopg2-binary>=2.9.9
